Created collectable keys/items, player raycasting and simple ui text
Created a collectable item prefab, located in the interactables folder (I'll want to change this location/folder name)

Created the user raycast function so we know what the center of the players screen is looking at. This includes items, keys and doors

Created a player UI Handler. This currently will assist in displaying information to the player about what item they are looking at