Created Simple Pause Function
Added the ability to pause the game.

Time.timescale is set to 0

the hud changes to the paused hud

The player action map swaps between "Player" and "UI"

Lastly added a check to see if we change input devices, and to show the mouse at relevant times (Changes to keyboard&mouse, game is paused. Changes to gamepad, hide mouse during pause)